remove data files
